refactor(channels): satisfy durable ingress gates
